fixed random, some exception fails

+++ matrix/sources/main.cpp
@@ -5,13 +5,12 @@
 
 int main ()
 {
-    int n;
+    size_t n;
     std::cin >> n;
 
     matrix::Matrix<int> matr (n, n);
 
     std::cin >> matr;
-    
     std::cout << matr.det () << std::endl;
 
     return 0;
